Move standalone logic out of shared code
