+++ src/coreclr/nativeaot/Runtime/gcenv.h
@@ -48,6 +48,8 @@
 #include "gcenv.inl"
 
 #include "stressLog.h"
+
+#ifndef SKIP_TRACING_DEFINITIONS
 #ifdef FEATURE_EVENT_TRACE
 
     #include "clretwallmain.h"
@@ -59,6 +61,7 @@
     #define ETW_EVENT_ENABLED(e,f) false
 
 #endif // FEATURE_EVENT_TRACE
+#endif //SKIP_TRACING_DEFINITIONS
 
 #define LOG(x)
