+++ src/coreclr/nativeaot/Runtime/gcheaputilities.h
@@ -108,4 +108,23 @@ private:
     GCHeapUtilities() = delete;
 };
 
+#ifndef DACCESS_COMPILE
+
+enum GC_LOAD_STATUS {
+    GC_LOAD_STATUS_BEFORE_START,
+    GC_LOAD_STATUS_START,
+    GC_LOAD_STATUS_DONE_LOAD,
+    GC_LOAD_STATUS_GET_VERSIONINFO,
+    GC_LOAD_STATUS_CALL_VERSIONINFO,
+    GC_LOAD_STATUS_DONE_VERSION_CHECK,
+    GC_LOAD_STATUS_GET_INITIALIZE,
+    GC_LOAD_STATUS_LOAD_COMPLETE
+};
+
+extern GC_LOAD_STATUS g_gc_load_status;
+
+extern VersionInfo g_gc_version_info;
+
+#endif //DACCESS_COMPILE
+
 #endif // _GCHEAPUTILITIES_H_
